Refactor the "free_userinfo()" function
This commit introduces a few style improvements.  It should be
considered a first step in improving the style within the module as
a whole.

First, we use "if (!userinfo)" instead of "if (userinfo == NULL)"
for two reasons:

 1. Accidental assignment is avoided.
 2. We no longer depend on the NULL symbol being defined.

Next, after checking that "userinfo" is not NULL, we can free the
inner character pointer without worrying if it's NULL.  This is
because "free()" does nothing when passed a NULL pointer[1].

Finally, casting to "(void *)" is not necessary here.  This is only
necessary when attempting to avoid the "-Wdiscarded-qualifiers"
compiler warning.  This warning won't be generated here because none
of the pointers involved are qualified with the "const" qualifier.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/