Use definitions from the "stdbool.h" header
It is generally preferable to use <stdbool.h> when you require
"true" and "false" macros.  Using this header allows for tighter
checking of types and values at compile-time.  It is also more in
line with the standards of modern C, having been introduced in C99.

Furthermore, using this header will allow the debugger to refer to
"false" and "true" values using their symbolic names rather than
their numeric values, 0 and 1.  This makes debugging easier and far
more natural.