Don't use pam environment in setcred
At least Arch runs pam_env from system-auth without the user_readenv
option, so ~/.pam_environment is ignored there. Therefore,
gpg-connect-agent was called with a mostly empty environment, breaking
the setup of users who override XDG_CONFIG_HOME or GNUPGHOME (e.g. #23).

I assume that in most relevant cases, we are either opening a new
session and should use the pam environment, or we are just
authenticating within an existing session, where the env is already set
up in the calling process and should be reused. Therefore, we now do not
use the pam environment during setcred anymore.

It may turn out to be necessary to add another option to make this
configurable.